Resolved class loading issues
When loading the abiquo jars in the classpath using the
'sys.path.extend(...)' method a classloader warning appeared
saying: "Could not load Finalizer in its own class loader".

The related class are part of the Guava library and have some
particularities that make this warnings appear when loaded by
the application classloader. If those classes are loaded by the
system classloader instead, there should be no problems, as explained
in the javadoc of the affected classes:
http://code.google.com/p/guava-libraries/source/browse/src/com/google/common/base/FinalizableReferenceQueue.java?r=f082dc3b19dfc77cf02f53ba0ca1b7a5ca5ee060

For the moment, I've created a wrapper shell script to export
the CLASSPATH variable in order to customize the system classloader
and make the dependencies be loaded in it. It would be interesting,
however, to investigate the Jynx (http://www.fiber-space.de/jynx/doc/)
library and the jclasspath module (http://www.fiber-space.de/jynx/doc/jclasspath.html)
to dynamically manipulate the system classpath. This requires Jython
2.5 or better and a little tunning of the Java access protection
policies in order to make protected methods in ClassLoaders accessible.
In this scenario, it might be better to export the environment variable
to make the project more portable.